Reestructuración arquitectónica: Migración a Supabase RPC, gestión avanzada de sesiones y optimización del flujo de reservas
* **Migración a Supabase:** Sustitución de Firebase Firestore por Supabase RPC y Realtime para la gestión de reservas, sedes y búsqueda de usuarios QuickPlay. Se implementaron funciones remotas (`get_reserva_completa`, `buscar_usuarios_quickplay`, `validar_hora_inicio_reserva`) para centralizar la lógica en el servidor.
* **Gestión de Sesiones y Auth:** Refactorización profunda de `AuthRepositoryImpl` y `CheckSessionUseCase`. Se integró el plugin `Auth` de Ktor con soporte para `BearerTokens` y renovación automática de tokens (`refreshTokens`) mediante un mutex. Ahora se gestiona la rotación de `refresh_token` de Supabase y la persistencia en `UserSessionManager`.
* **Lógica de Reservas:**
    * Rediseño de `ReservaRepository` para utilizar flujos reactivos basados en canales de Supabase Realtime.
    * Actualización de tipos de datos: migración de IDs de `String` a `Int` en los DTOs de reserva, juegos y logs de monedas.
    * Implementación de `ValidarHoraReservaUseCase` para verificar la viabilidad de inicio de partidas desde el servidor.
* **Seguridad y Registro:** Adición de `PasswordRequirements` y el componente `PasswordRequirementsIndicator` en el registro de usuarios para validar longitud, mayúsculas, números y caracteres especiales en tiempo real.
* **UI y UX:**
    * **Reservas:** Mejora del `ReservasModel` con soporte para Pull-to-Refresh manual y filtrado optimizado por estado (Activas, Pendientes, Historial).
    * **Aceptar Reto:** Rediseño de la pantalla de aceptación de retos, simplificando las acciones tras el éxito y mejorando la visualización del desglose de pagos.
    * **Moneda:** Optimización del flujo de compra de monedas, asegurando el registro del log de auditoría antes de la redirección a WhatsApp.
* **Infraestructura y Utilidades:**
    * Mejora de `DateTimeFormatter` para manejar conversiones ISO con offset de zona horaria (`America/Lima`).
    * Optimización de `UserSessionManager` con limpieza asíncrona de tablas SQLite (`Usuario`, `Inventario`, `Horario`) al cerrar sesión.
    * Limpieza de animaciones de batalla, eliminando parámetros y botones redundantes en `BattleWinnerAnimation` y `BattleLoserAnimation`.

+++ composeApp/src/commonMain/kotlin/pe/breaker/dkaviplay/data/repository/JuegoRepositoryImpl.kt
@@ -67,10 +67,10 @@ class JuegoRepositoryImpl(
         }
 
         val requestBody = ActualizarPuntosRequestDTO(
-            juegoId = reserva.juegoUid,
+            juegoId = reserva.juegoUid.toInt(),
             partidas = partidas,
             tipoJuego = reserva.tipoJuego,
-            reservaId = reserva.reservaUid,
+            reservaId = reserva.reservaUid.toInt(),
             contrincanteUid = contrincanteUid,
             sender = sessionManager.getCurrentUsuario()?.usuario
         )
@@ -86,8 +86,8 @@ class JuegoRepositoryImpl(
 
     override suspend fun cerrarAcuerdoMutuo(reserva: Reserva): Result<String> {
         val requestBody = CerrarJuegoMutuoRequestDTO(
-            reservaId = reserva.reservaUid,
-            juegoId = reserva.juegoUid
+            reservaId = reserva.reservaUid.toInt(),
+            juegoId = reserva.juegoUid.toInt()
         )
         val response =
             httpClient.post("${ConstatesCloud.URLBASE}apiPublic/public/juego/confirmJuegoService") {
